requirements.txt
absl-py==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
diffusers==0.27.2
filelock==3.13.4
fsspec==2024.3.1
grpcio==1.62.2
huggingface-hub==0.22.2
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.3
joblib==1.4.0
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
timm==0.9.16
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
